feat: 커스텀 질문 validatioin 추가 수정 부분의 오류 수정할 것 request param에 validation 추가하는 방법 알아볼 것

+++ src/main/java/com/example/amattang/controller/CustomQuestionController.java
@@ -1,6 +1,9 @@
 package com.example.amattang.controller;
 
 
+import com.example.amattang.controller.ValidationGroup.generalGroup;
+import com.example.amattang.controller.ValidationGroup.saveGroup;
+import com.example.amattang.controller.ValidationGroup.updateGroup;
 import com.example.amattang.domain.user.User;
 import com.example.amattang.domain.user.UserRepository;
 import com.example.amattang.payload.reponse.CustomCheckListResponseDto;
@@ -12,18 +15,18 @@ import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Arrays;
+import javax.validation.Valid;
+import javax.validation.constraints.Size;
 import java.util.List;
 import java.util.NoSuchElementException;
 
 import static com.example.amattang.payload.reponse.ResponseMessage.*;
 import static com.example.amattang.payload.reponse.ResponseUtil.succes;
 import static org.springframework.http.HttpStatus.CREATED;
-import static org.springframework.http.HttpStatus.OK;
 
 @RestController
 @RequiredArgsConstructor
@@ -50,7 +53,7 @@ public class CustomQuestionController {
             @ApiImplicitParam(name = "CustomQuestionRequestDto")
     })
     @PostMapping
-    public ResponseEntity<?> doc(@CurrentUser UserPrincipal userPrincipal, @PathVariable("checkListId") Long checkListId, @RequestBody CustomRequestDto dto) {
+    public ResponseEntity<?> doc(@CurrentUser UserPrincipal userPrincipal, @PathVariable("checkListId") Long checkListId, @RequestBody @Validated({saveGroup.class, generalGroup.class}) CustomRequestDto dto) {
         User user = userRepository.findById(userPrincipal.getId()).orElseThrow(() -> new NoSuchElementException("사용자를 찾을 수 없습니다."));
         questionService.saveCategoryAndAnswer(user, checkListId, dto);
         return succes(CREATED, CREATE_CUSTOM_QUESTION);
@@ -62,7 +65,7 @@ public class CustomQuestionController {
             @ApiImplicitParam(name = "CustomQuestionRequestDto")
     })
     @PutMapping
-    public ResponseEntity<?> doc2(@CurrentUser UserPrincipal userPrincipal, @PathVariable("checkListId") Long checkListId, @RequestBody CustomRequestDto dto) {
+    public ResponseEntity<?> doc2(@CurrentUser UserPrincipal userPrincipal, @PathVariable("checkListId") Long checkListId, @RequestBody @Validated({updateGroup.class, generalGroup.class}) CustomRequestDto dto) {
         User user = userRepository.findById(userPrincipal.getId()).orElseThrow(() -> new NoSuchElementException("사용자를 찾을 수 없습니다."));
         questionService.updateCategoryAndAnswer(user, checkListId, dto);
         return succes(CREATED, UPDATE_CUSTOM_QUESTION);
@@ -76,22 +79,10 @@ public class CustomQuestionController {
     })
     @DeleteMapping
     public ResponseEntity<?> doc11(@CurrentUser UserPrincipal userPrincipal, @PathVariable("checkListId") Long checkListId, @RequestParam("categoryId") List<Long> categoryIds) {
+        if (categoryIds.size() == 0) throw new IllegalArgumentException("카테고리 아이디 리스트는 최소 1개 이상의 아이디를 포함해야 합니다.");
         User user = userRepository.findById(userPrincipal.getId()).orElseThrow(() -> new NoSuchElementException("사용자를 찾을 수 없습니다."));
         questionService.deleteCategories(user, checkListId, categoryIds);
         return succes(CREATED, DELETE_CUSTOM_CATEGORY);
-    }
-
-    @ApiOperation(value = "3-5. 카테고리의 상세 항목 삭제")
-    @ApiImplicitParams({
-            @ApiImplicitParam(name = "checkListId", value = "체크리스트 아이디", paramType = "path"),
-            @ApiImplicitParam(name = "id", value = "커스텀 카테고리의 하위 항목 아이디 리스트", example = "3", required = true, paramType = "params")
-    })
-    @DeleteMapping("/question/{categoryId}")
-    public ResponseEntity<?> doc4(@CurrentUser UserPrincipal userPrincipal, @PathVariable("checkListId") Long checkListId,
-                                  @PathVariable("categoryId") Long categoryId, @RequestParam("id") List<Long> id) {
-        User user = userRepository.findById(userPrincipal.getId()).orElseThrow(() -> new NoSuchElementException("사용자를 찾을 수 없습니다."));
-        questionService.deleteCategoryItems(user, checkListId, categoryId, id);
-        return succes(OK, DELETE_CUSTOM_QUESTION);
     }
 
 }
